# server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/admin/login": {
      "post": {
        "operationId": "MainController_login",
        "summary": "用户登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/logout": {
      "post": {
        "operationId": "MainController_logout",
        "summary": "退出登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/register": {
      "post": {
        "operationId": "MainController_register",
        "summary": "用户注册",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/registerUser": {
      "get": {
        "operationId": "MainController_registerUser",
        "summary": "账号自助-是否开启用户注册功能",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/captchaImage": {
      "get": {
        "operationId": "MainController_captchaImage",
        "summary": "获取验证图片",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/getInfo": {
      "get": {
        "operationId": "MainController_getInfo",
        "summary": "用户信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/admin/getRouters": {
      "get": {
        "operationId": "MainController_getRouters",
        "summary": "路由信息",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        },
        "required": [
          "userName",
          "password",
          "uuid"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        },
        "required": [
          "userName",
          "password",
          "uuid"
        ]
      },
...
    }
  }
}